% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.obj.ee.R
\name{eval.obj.ee}
\alias{eval.obj.ee}
\title{Function for the evaluation of the enrichment of subnetwork x in two ranked lists of differentially expressed genes}
\usage{
eval.obj.ee(x = NULL, eval_obj.p = NULL)
}
\arguments{
\item{x}{a vector of ids corresponding to the nodes of a subnetwork}

\item{eval_obj.p}{list with elements \code{exprs1} and \code{exprs2}, two sorted and named vectors of gene expression variations}
}
\value{
numeric vector with the log10 of p-values assessing the enrichment of subnetwork x in differentially expressed genes at the top of \code{exprs1} and \code{exprs2} vectors.
}
\description{
Function for the evaluation of the enrichment of subnetwork x in two ranked lists of differentially expressed genes
}
\author{
\packageAuthor{mosub}
}
