% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosub_data.R
\docType{data}
\name{ge2}
\alias{ge2}
\title{Gene expression data included in package mosub}
\format{
vector of gene log2-fold changes between breast carcinoma and normal samples; \code{names(ge2)} are Entrez Gene ids.
}
\source{
The Cancer Genoma Atlas http://tcga-data.nci.nih.gov/tcga/.
}
\usage{
ge2
}
\description{
Gene expression data included in package mosub
}
\keyword{datasets}
