% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosub_data.R
\docType{data}
\name{interactions}
\alias{interactions}
\title{Protein-protein interaction and gene expression data included in package mosub.}
\format{
\code{data.table} containing human protein-protein associations:
\describe{
  \item{a}{Entrez Gene id;}
  \item{b}{Entrez Gene id;}
  \item{score}{confidence score.}
}
}
\source{
STRING database
}
\usage{
interactions
}
\description{
Protein-protein interaction and gene expression data included in package mosub.
}
\references{
Szklarczyk, D., et al., 2011. The STRING database in 2011: functional interaction networks of proteins, globally integrated and scored. Nucleic Acids Res. 39: D561-8.
}
\keyword{datasets}
