% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplane.R
\name{kplane}
\alias{kplane}
\title{k-Plane Clustering Algorithm}
\usage{
kplane(X, K = 2, iter = 496, print.progress = TRUE)
}
\arguments{
\item{X}{an \eqn{(n\times p) data matrix.}}

\item{K}{the number of clusters.}

\item{iter}{the number of EM-type updating.}

\item{print.progress}{a logical; \code{TRUE} to show completion of iterations by 10, \code{FALSE} otherwise.}
}
\value{
a list whose elements are also lists of following elements: \describe{
\item{cluster}{length-\code{n} vector of cluster label.}
\item{w}{an \eqn{(p\times K)} matrix of orthonormal columns.}
\item{gamma}{a length-\eqn{K} vector of projection values.}
}
}
\description{
It is one of the first method proposed in 2000 for fitting \eqn{K} planes by EM-like algorithm 
that updates parametrization of a plane and assigning data points to the nearest plane in an alternating manner.
Empirically, its performance is not competitive against newer methods but we include this as a test benchmark 
as well as it is simple and fast enough.
}
\examples{
## generate a toy example of two plane components
set.seed(18)
tester = gen.LP2(n=100, nl=0, np=2)
data   = tester$data
label  = tester$class

## do PCA for data reduction
proj = base::eigen(stats::cov(data))$vectors[,1:2]
dat2 = data\%*\%proj

## run k-plane algorithm with K=2, 3, and 4
kplane2 = kplane(data, K=2)
kplane3 = kplane(data, K=3)
kplane4 = kplane(data, K=4)

## extract clustering
finc2 = kplane2$cluster
finc3 = kplane3$cluster
finc4 = kplane4$cluster

## visualize
opar <- par(mfrow=c(3,4))
plot(dat2[,1],dat2[,2],pch=19,cex=0.3,col=finc2+1,main="K=2:PCA")
plot(data[,1],data[,2],pch=19,cex=0.3,col=finc2+1,main="K=2:Axis(1,2)")
plot(data[,1],data[,3],pch=19,cex=0.3,col=finc2+1,main="K=2:Axis(1,3)")
plot(data[,2],data[,3],pch=19,cex=0.3,col=finc2+1,main="K=2:Axis(2,3)")

plot(dat2[,1],dat2[,2],pch=19,cex=0.3,col=finc3+1,main="K=3:PCA")
plot(data[,1],data[,2],pch=19,cex=0.3,col=finc3+1,main="K=3:Axis(1,2)")
plot(data[,1],data[,3],pch=19,cex=0.3,col=finc3+1,main="K=3:Axis(1,3)")
plot(data[,2],data[,3],pch=19,cex=0.3,col=finc3+1,main="K=3:Axis(2,3)")

plot(dat2[,1],dat2[,2],pch=19,cex=0.3,col=finc4+1,main="K=4:PCA")
plot(data[,1],data[,2],pch=19,cex=0.3,col=finc4+1,main="K=4:Axis(1,2)")
plot(data[,1],data[,3],pch=19,cex=0.3,col=finc4+1,main="K=4:Axis(1,3)")
plot(data[,2],data[,3],pch=19,cex=0.3,col=finc4+1,main="K=4:Axis(2,3)")
par(opar)

}
\references{
\insertRef{bradley_k-plane_2000}{mosub}
}
