% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn.sim.R
\name{mn.sim}
\alias{mn.sim}
\title{Calculation of the pair-wise similarity of a population of individuals}
\usage{
mn.sim(x = NULL)
}
\arguments{
\item{x}{population of individuals defined by \code{mosub_find}}
}
\value{
matrix of similarities
}
\description{
Calculation of the pair-wise similarity of a population of individuals
}
\author{
\packageAuthor{mosub}
}
