% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosub-package.R
\docType{package}
\name{mosub}
\alias{mosub}
\alias{mosub-package}
\title{Bayesian Mixture of Subspaces of Different Dimensions}
\description{
Bayesian model for inferring mixtures of subspaces of different dimensions is presented.
}
