% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmpred.R
\name{msmpred}
\alias{msmpred}
\title{Predict Class Labels from Mixtures of Subspaces}
\usage{
msmpred(X, msmoutput, print.progress = TRUE)
}
\arguments{
\item{X}{an \eqn{(n\times p) data matrix.}}

\item{msmoutput}{output of \code{msm} function.}

\item{print.progress}{a logical; \code{TRUE} to show completion of iterations by 10, \code{FALSE} otherwise.}
}
\value{
a length-\code{n} vector of class labels.
}
\description{
Once the run from main function \code{msm} is done, \code{msmpred} predicts 
the class label for the given data. In this function, there is no need for an user 
to modify any output. It simply takes a returned object from \code{msm} function, 
which is a list of lists, and new data.
}
\examples{
## generate a toy example
set.seed(10)
alldat = gen.LP(n=500, K=2, iso.var=0.1)

## separate as train/test data
id.train = sample(1:1000, 800, replace=FALSE)
id.test  = setdiff(1:1000, id.train)

train.dat = alldat$data[id.train,]
train.lab = alldat$class[id.train]

test.dat = alldat$data[id.test,]
test.lab = alldat$class[id.test]

## run MSM algorithm with K=2
maxiter   = 10000
output2   = msm(train.dat, K=2, iter=maxiter)
test.pred = msmpred(test.dat, output2)

## visualize with axis Y and Z for the test data
tX = test.dat[,2]
tY = test.dat[,3]

opar <- par(mfrow=c(1,2))
plot(tX,tY,col=test.lab, pch=19,cex=0.5,main="true label")
plot(tX,tY,col=test.pred,pch=19,cex=0.5,main="predicted label")
par(opar)

}
