% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.hypervolume.R
\name{pareto.hypervolume}
\alias{pareto.hypervolume}
\title{Monte-Carlo Calculation of Pareto hypervolume}
\usage{
pareto.hypervolume(P = NULL, r = c(1, 1), N = 10000)
}
\arguments{
\item{P}{matrix of objective values \(columns\) for a set of optimal solutions \(rows\)}

\item{r}{vector of maximum values of optimal solutions}

\item{N}{number of permutations}
}
\value{
numeric value, the hypervolume
}
\description{
Monte-Carlo Calculation of Pareto hypervolume
}
\author{
\packageAuthor{mosub}
}
