% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.normalize.R
\name{pareto.normalize}
\alias{pareto.normalize}
\title{Normalization of a pareto front}
\usage{
pareto.normalize(x = NULL)
}
\arguments{
\item{x}{n-by-m matrix with n points expressed with m coordinates}
}
\value{
scaled matrix
}
\description{
Normalization of a pareto front
}
\details{
This function scales the Pareto front in [0, 1], according to the formula: y = (x - min(x)) / (max(x) - min(x))
}
\author{
\packageAuthor{mosub}
}
