% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.postProcess.R
\name{population.postProcess}
\alias{population.postProcess}
\title{Post-processing of optimal subnetworks}
\usage{
population.postProcess(population = NULL, interactions = NULL)
}
\arguments{
\item{population}{list of individuals found by \code{mosub_find}}

\item{interactions}{data.frame with interactions}
}
\value{
a list that contains the population and igraphs objects associated with it
}
\description{
Post-processing of optimal subnetworks
}
\author{
\packageAuthor{mosub}
}
