% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.rnd.R
\name{population.rnd}
\alias{population.rnd}
\title{Random Generation of an Initial Population of Connected Subnetworks}
\usage{
population.rnd(
  pop = NULL,
  M = NULL,
  seeds = NULL,
  mn.size.max = 30,
  interactions = NULL,
  mn.size.min = 3,
  max.new.nodes = NULL,
  parallel = FALSE,
  mc.cores = mc.cores
)
}
\arguments{
\item{pop}{number of individuals}

\item{M}{number of objectives}

\item{seeds}{set of ids to be used as seeds in the generation of connected subnetworks}

\item{mn.size.max}{maximum size of subnetwork}

\item{interactions}{data.frame with interactions}

\item{mn.size.min}{minimum size of subnetwork}

\item{max.new.nodes}{maximum number of neighbors that will be added at each iteration of the algorithm for the generation of connected subnetworks}

\item{parallel}{TRUE/FALSE value that specifies wether to use parallel computations}

\item{mc.cores}{number of cores}
}
\value{
a population of individuals
}
\description{
Random Generation of an Initial Population of Connected Subnetworks
}
\author{
Ettore Mosca
}
