% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosub_data.R
\docType{data}
\name{res.10}
\alias{res.10}
\title{Output of mosub_find after 10 iterations}
\format{
list with elements \code{individuals}, \code{objectives}, \code{rank} and \code{cd}.
}
\usage{
res.10
}
\description{
Output of mosub_find after 10 iterations
}
\keyword{datasets}
