% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosub_data.R
\docType{data}
\name{res.500}
\alias{res.500}
\title{Output of mosub_find after 500 iterations}
\format{
list with elements \code{individuals}, \code{objectives}, \code{rank}, \code{cd} and \code{igraphs}.
}
\usage{
res.500
}
\description{
Output of mosub_find after 500 iterations
}
\keyword{datasets}
