% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.LP.R
\name{gen.LP}
\alias{gen.LP}
\title{Generate Line and Plane Example}
\usage{
gen.LP(n = 100, K = 2, iso.var = 0.1)
}
\arguments{
\item{n}{the number of data points for each line and plane.}

\item{K}{the number of mixture components.}

\item{iso.var}{degree of isotropic variance.}
}
\value{
a named list containing:\describe{
\item{data}{an \eqn{(2n\times 3)} data matrix.}
\item{class}{length-\eqn{2n} vector for class label.}
}
}
\description{
This function generates a toy example of 'line and plane' data in \eqn{R^3} that 
data are generated from a mixture of lines (one-dimensional) planes (two-dimensional).
When \code{K=2}, it gives one line and one plane. For \code{K>2}, it gives one line, 
one plane, and \code{K-2} componenets are randomly chosen to be either a line or a plane.
}
\examples{
## test for visualization
set.seed(10)
tester = gen.LP(n=100, K=2, iso.var=0.1)
data   = tester$data
label  = tester$class

## do PCA for data reduction
proj = base::eigen(stats::cov(data))$vectors[,1:2]
dat2 = data\%*\%proj

## visualize
opar <- par(mfrow=c(2,2))
plot(dat2[,1],dat2[,2],pch=19,cex=0.5,col=label,main="PCA")
plot(data[,1],data[,2],pch=19,cex=0.5,col=label,main="Axis 1 vs 2")
plot(data[,1],data[,3],pch=19,cex=0.5,col=label,main="Axis 1 vs 3")
plot(data[,2],data[,3],pch=19,cex=0.5,col=label,main="Axis 2 vs 3")
par(opar)

\dontrun{
## visualize in 3d
x11()
scatterplot3d::scatterplot3d(x=data, pch=19, cex.symbols=0.5, color=label)
}

}
