% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.R
\name{msm}
\alias{msm}
\title{Mixtures of Varying-Dimensional Subspaces Model}
\usage{
msm(
  X,
  K = 2,
  iter = 496,
  prop.var = 1,
  temperature = 1e-06,
  print.progress = TRUE
)
}
\arguments{
\item{X}{an \eqn{(n\times p) data matrix.}}

\item{K}{the number of mixtures.}

\item{iter}{the number of MCMC runs.}

\item{prop.var}{proposal variance parameter.}

\item{temperature}{temperature value for Gibbs posterior.}

\item{print.progress}{a logical; \code{TRUE} to show completion of iterations by 10, \code{FALSE} otherwise.}
}
\value{
a length-\code{iter} list whose elements are also lists of following elements: \describe{
\item{P}{length-\code{K} list of projection matrices.}
\item{U}{length-\code{K} list of orthonormal basis.}
\item{theta}{length-\code{K} list of center locations of each mixture.}
\item{cluster}{length-\code{n} vector of cluster label.}
}
}
\description{
\code{msm} is a Bayesian model inferring mixtures of subspaces that are of possibly different dimensions. 
For simplicity, this function returns only a handful of information that are most important in 
representing the mixture model, including projection, location, and hard assignment parameters.
}
\examples{
## generate a toy example
set.seed(10)
tester = gen.LP(n=100, K=2, iso.var=0.1)
data   = tester$data
label  = tester$class

## do PCA for data reduction
proj = base::eigen(stats::cov(data))$vectors[,1:2]
dat2 = data\%*\%proj

## run MSM algorithm with K=2, 3, and 4
maxiter = 5000
output2 = msm(data, K=2, iter=maxiter)
output3 = msm(data, K=3, iter=maxiter)
output4 = msm(data, K=4, iter=maxiter)

## extract final clustering information
finc2 = output2[[maxiter]]$cluster
finc3 = output3[[maxiter]]$cluster
finc4 = output4[[maxiter]]$cluster

## visualize
opar <- par(mfrow=c(3,4))
plot(dat2[,1],dat2[,2],pch=19,cex=0.5,col=finc2,main="K=2:PCA")
plot(data[,1],data[,2],pch=19,cex=0.5,col=finc2,main="K=2:Axis(1,2)")
plot(data[,1],data[,3],pch=19,cex=0.5,col=finc2,main="K=2:Axis(1,3)")
plot(data[,2],data[,3],pch=19,cex=0.5,col=finc2,main="K=2:Axis(2,3)")

plot(dat2[,1],dat2[,2],pch=19,cex=0.5,col=finc3,main="K=3:PCA")
plot(data[,1],data[,2],pch=19,cex=0.5,col=finc3,main="K=3:Axis(1,2)")
plot(data[,1],data[,3],pch=19,cex=0.5,col=finc3,main="K=3:Axis(1,3)")
plot(data[,2],data[,3],pch=19,cex=0.5,col=finc3,main="K=3:Axis(2,3)")

plot(dat2[,1],dat2[,2],pch=19,cex=0.5,col=finc4,main="K=4:PCA")
plot(data[,1],data[,2],pch=19,cex=0.5,col=finc4,main="K=4:Axis(1,2)")
plot(data[,1],data[,3],pch=19,cex=0.5,col=finc4,main="K=4:Axis(1,3)")
plot(data[,2],data[,3],pch=19,cex=0.5,col=finc4,main="K=4:Axis(2,3)")
par(opar)

}
\references{
\insertRef{thomas_learning_2015}{mosub}
}
